requirements.txt
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
openai==1.44.0
pydantic==2.9.1
pydantic_core==2.23.3
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
telebot==0.0.5
tqdm==4.66.5
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
